Move QR code implementations to module
