+++ qr-code/build.gradle.kts
@@ -48,6 +48,10 @@ dependencies {
     implementation(libs.zxingAndroidEmbedded)
     implementation(libs.mlkit.barcodescanning)
     implementation(libs.androidxLifecycleRuntimeKtx)
+    implementation(libs.androidMaterial)
+    implementation(libs.cameraxView)
+    implementation(libs.mlkit.barcodescanning)
+    implementation(libs.camera.mlkit.vision)
 
     testImplementation(libs.androidxTestExtJunit)
     testImplementation(libs.hamcrest)
