+++ qr-code/src/main/java/org/odk/collect/qrcode/MlKitBarcodeScannerViewFactory.kt
@@ -1,4 +1,4 @@
-package org.odk.collect.android.fragments
+package org.odk.collect.qrcode
 
 import android.annotation.SuppressLint
 import android.app.Activity
@@ -16,9 +16,7 @@ import com.google.android.gms.common.moduleinstall.ModuleInstall
 import com.google.mlkit.vision.barcode.BarcodeScannerOptions
 import com.google.mlkit.vision.barcode.BarcodeScanning
 import com.google.mlkit.vision.barcode.common.Barcode
-import org.odk.collect.android.databinding.MlkitBarcodeScannerLayoutBinding
-import org.odk.collect.qrcode.BarcodeScannerView
-import org.odk.collect.qrcode.BarcodeScannerViewContainer
+import org.odk.collect.qrcode.databinding.MlkitBarcodeScannerLayoutBinding
 
 @SuppressLint("ViewConstructor")
 private class MlKitBarcodeScannerView(
