+++ qr-code/src/main/java/org/odk/collect/qrcode/ZxingBarcodeScannerViewFactory.kt
@@ -1,4 +1,4 @@
-package org.odk.collect.android.fragments
+package org.odk.collect.qrcode
 
 import android.annotation.SuppressLint
 import android.app.Activity
@@ -7,11 +7,8 @@ import androidx.lifecycle.DefaultLifecycleObserver
 import androidx.lifecycle.LifecycleOwner
 import com.google.zxing.integration.android.IntentIntegrator
 import com.journeyapps.barcodescanner.camera.CameraSettings
-import org.odk.collect.android.databinding.ZxingBarcodeScannerLayoutBinding
-import org.odk.collect.android.utilities.CodeCaptureManagerFactory
 import org.odk.collect.androidshared.system.CameraUtils
-import org.odk.collect.qrcode.BarcodeScannerView
-import org.odk.collect.qrcode.BarcodeScannerViewContainer
+import org.odk.collect.qrcode.databinding.ZxingBarcodeScannerLayoutBinding
 
 @SuppressLint("ViewConstructor")
 private class ZxingBarcodeScannerView(
