+++ qr-code/src/main/res/layout/zxing_barcode_scanner_layout.xml
